requirements.txt
google-adk
load-dotenv
vertexai